requirements.txt
flask==3.0.3
pandas==2.2.2
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2